exception suggestion2
I made exception by making the case. For example, if the M head for north and it locates at x=0, it will read only the front sensor, and right sensor.
But, if this case, the M will continue moving, but it will not detect front wall. Next step is that I will make new variable number and set the token of stopping sensoring.