success_continuous_adachi&try_slalom
I succeeded in making code of continuous_adachi_method and tried making code of slalom run.
In addition, I converted the trapezoid groups into the code which do not have motor's lock or pwm on-off.
slalom run should be adjusted from the point of the velocity and angle velocity, and so on. Or should cut off the wall_control_flag.